// frontend/src/app/api/model/boxesOverviewItem.ts
/**
 * MemoQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BoxesOverviewItem { 
    boxNumber?: number;
    cardCount?: number;
    isToday?: boolean;
}