// frontend/src/app/api/model/lastSessionSummary.ts
/**
 * MemoQuiz API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LastSessionSummary { 
    reviewedCards?: number;
    goodAnswers?: number;
    successRate?: number;
    startedAt?: string;
    dayIndex?: number;
}